Solucion error en editar info del usuario

+++ src/main/java/com/Arris/controllers/dto/UsuarioRegistroDTO.java
@@ -69,7 +69,7 @@ public class UsuarioRegistroDTO {
         return idUsuario;
     }
 
-    public void setIdUsuario(Integer idUsuario) {
+    public void setIdUsuario(long idUsuario) {
         this.idUsuario = idUsuario;
     }
